refactor(core): reorganize navbar imports
