+++ src/app/core/navbar/navbar-item/navbar-item.component.ts
@@ -4,7 +4,7 @@ import { RouterLink, RouterLinkActive } from '@angular/router';
 import { FaIconComponent } from '@fortawesome/angular-fontawesome';
 import { NgbDropdownToggle } from '@ng-bootstrap/ng-bootstrap';
 
-import { NavbarItem } from '../models/navbar.model';
+import { NavbarItem } from '../navbar.model';
 
 /**
  * The NavbarItem component.
